+++ app/src/main/java/com/henryhiles/qweather/domain/weather/WeatherData.kt
@@ -4,7 +4,7 @@ import java.time.LocalDateTime
 
 data class WeatherData(
     val time: LocalDateTime,
-    val temperatureCelsius: Double,
+    val temperatureCelsius: Int,
     val pressure: Double,
     val windSpeed: Double,
     val humidity: Double,
